+++ src/lib/button/Button.svelte
@@ -26,6 +26,8 @@
 
 	const uid = $props.id()
 
+	const isLink = $derived(attributes.href != null && !disabled && !loading)
+
 	const tooltipId = $derived(title && !disabled && !loading ? uid : undefined)
 
 	let pressed = $state(false)
@@ -72,7 +74,7 @@
 	{/if}
 {/snippet}
 
-{#if 'href' in attributes && !disabled && !loading}
+{#if isLink}
 	<a
 		{...attributes}
 		onclick={(event) => {